\name{cov}
\alias{cov}
\docType{data}
\title{ 
  COV
}
\description{
Example covariates data frame.
}
\usage{data("cov")}
\format{
  A data frame with 600 rows and 5 columns of NRVATGFLMM.
\details{
A data frame containing the covariates information. The first two columns are needed to be named
“ped” and “person”, which are used to match subjects to their data in the pedigree data frame.
}
\seealso{
\code{\link{Ped}}, \code{\link{geno}}, \code{\link{snpPos}}, and \code{\link{Kin2}}
}
