\name{NRVATGFLMM-package}
\alias{NRVATGFLMM-package}
\alias{NRVATGFLMM}
\docType{package}
\title{
Novel Rare Variants Association Test and Generalized Functional Linear Mixed Models for Binary Traits in Family-Based Designs via Copulas
}
\description{
An R package for performing a family-based association test for rare variants using marginal logistic (NRVAT) and generalized functional linear mixed models (CBGF) for gene-based association testing of dichotomous traits implying a Gaussian Copula. The latter is employed to model the dependence between relatives (Dossa et al, 2022).
}
\author{
DOSSA H. Roland G., Lakhal-Chaieb Lajmi, Oualkacha Karim
}

%\references{
%~~ Literature or other references for background information ~~
%}
